Simplify pass: dedup the @TargetStream contract javadoc, move @TargetStream to occurrent-annotations
Two quality fixes from a /simplify review of the command-dispatch feature:

- The "exactly one property" contract was restated in full on both @TargetStream
  and AnnotationStreamIdResolver. State it once on @TargetStream and have the
  resolver reference it.
- Move @TargetStream from the command-dispatch module to occurrent-annotations,
  where every other Occurrent marker annotation lives (@DcbTag, @Saga, @Projection,
  @Subscription, @StreamId, ...). command-dispatch-annotation now depends on
  occurrent-annotations, mirroring how dcb-annotation depends on it for @DcbTag.

The ~150-line reflection-scanner duplication between AnnotationStreamIdResolver and
AnnotationTagGenerator is left for a separate follow-up, since deduping it edits the
existing AnnotationTagGenerator and needs a shared-module-home decision.
