+++ changelog.md
@@ -2,7 +2,7 @@
 
 #### Changes
 
-* Added a reusable command-dispatch capability, extracted from the Saga DSL so a command producer other than a saga can use it. `CommandDispatcher<C>`, the producer-facing port that issues a command (previously in the saga module), now lives in the new `occurrent-command-dispatch` module together with `StreamIdResolver<C>`, a named command-to-stream-id function that is the stream-side counterpart of the DCB `TagGenerator`, and `@TargetStream`, which marks the command member whose value is the target stream id. `occurrent-command-dispatch-annotation` adds `AnnotationStreamIdResolver`, a reflection-based `StreamIdResolver` that reads `@TargetStream` (or a custom marker annotation) off a command, mirroring the DCB `AnnotationTagGenerator`/`@DcbTag`. The blocking Spring Boot starter contributes a default `AnnotationStreamIdResolver` bean, overridable with your own `StreamIdResolver`. This is stream-side only, since DCB routes a command to its append boundary through tags (`@DcbTag`/`TagGenerator`, `DcbDecider.criteriaFor`) rather than a stream id. `CommandDispatchers.decider(...)` stays in the saga module, now taking a `StreamIdResolver` in place of a bare `Function`.
+* Added a reusable command-dispatch capability, extracted from the Saga DSL so a command producer other than a saga can use it. `CommandDispatcher<C>`, the producer-facing port that issues a command (previously in the saga module), now lives in the new `occurrent-command-dispatch` module together with `StreamIdResolver<C>`, a named command-to-stream-id function that is the stream-side counterpart of the DCB `TagGenerator`. The new `@TargetStream` annotation, which marks the command member whose value is the target stream id, joins the other Occurrent markers in `occurrent-annotations`. `occurrent-command-dispatch-annotation` adds `AnnotationStreamIdResolver`, a reflection-based `StreamIdResolver` that reads `@TargetStream` (or a custom marker annotation) off a command, mirroring the DCB `AnnotationTagGenerator`/`@DcbTag`. The blocking Spring Boot starter contributes a default `AnnotationStreamIdResolver` bean, overridable with your own `StreamIdResolver`. This is stream-side only, since DCB routes a command to its append boundary through tags (`@DcbTag`/`TagGenerator`, `DcbDecider.criteriaFor`) rather than a stream id. `CommandDispatchers.decider(...)` stays in the saga module, now taking a `StreamIdResolver` in place of a bare `Function`.
 * A subscription handler can now receive the stream id or stream version directly through the new `@StreamId` and `@StreamVersion` parameter annotations, so `void on(MyEvent event, @StreamId String streamId, @StreamVersion long streamVersion)` works without declaring an `EventMetadata` parameter and calling `getStreamId()`/`getStreamVersion()`. The annotations may appear in any order alongside the event and an optional `EventMetadata` parameter, on `@Subscription`, `@StreamSubscription`, and `@SynchronousSubscription`, for both the blocking and reactive stacks. `@StreamId` binds a `String`, `@StreamVersion` a `long` or `Long`. They are rejected at startup on `@DcbSubscription`, whose stream id/version are internal partition values. On the capability-agnostic `@Subscription` a DCB-appended event exposes the internal partition id and per-partition counter, the same values `EventMetadata` already gives there. The existing `EventMetadata` parameter form is unchanged.
 * The MongoDB change stream backing a subscription can now be tuned for throughput and latency. `NativeMongoSubscriptionModelConfig` gains `batchSize(int)` and `maxAwaitTime(Duration)`, and `SpringMongoSubscriptionModelConfig` gains `maxAwaitTime(Duration)`. A larger `batchSize` reduces server round-trips (throughput), and `maxAwaitTime` bounds how long the server holds an idle `getMore` before returning a (possibly empty) batch (latency versus resource usage). Both options are opt-in and default to unset, so the driver/server defaults still apply and no existing subscription changes behavior on upgrade. The options differ per model because Spring Data's abstractions do not expose the full set: the native driver's `ChangeStreamIterable` supports both, while the Spring blocking model reads the change stream through Spring's `MessageListenerContainer`, whose `ChangeStreamRequest`/`ChangeStreamRequestOptions` API carries a `maxAwaitTime` but no batch size (`ChangeStreamTask` never applies one), so `batchSize` is only available on the native model. `ReactorMongoSubscriptionModel` exposes neither yet, because Spring Data's `ReactiveMongoTemplate.changeStream` and its `ChangeStreamOptions` carry neither, and adding them means driving the raw reactive driver directly (deferred as a follow-up). Resolves [#173](https://github.com/johanhaleby/occurrent/issues/173). Rationale in [ADR 67](doc/architecture/decisions/0067-configurable-change-stream-batchsize-and-maxawaittime.md).
 * Nicer Kotlin ergonomics for building DCB query criteria from event classes. `queries.criteria()` (and `subscriptions.criteria()`) now lets you write `type<OrderPlaced>()` and `types<OrderPlaced, OrderCancelled>()` with the base event type inferred, or the `KClass` forms `type(OrderPlaced::class)` and `types(OrderPlaced::class, OrderCancelled::class)`, instead of `::class.java`. You can also reuse a shared boundary and give it query-specific types with `criteria(boundary).types<...>()`, where `boundary` is a `DcbCriterion` (a single alternative, so it stays within the OR-of-items model, [ADR 32](doc/architecture/decisions/0032-fluent-dcb-query-construction.md)). This is a Kotlin-only convenience: `DcbCriteriaBuilder` became a Kotlin class so the reified helpers can infer the base type as members, and its Java surface is unchanged. The old `typeOf<T, E>()` is deprecated in favor of `type<T>()` (the base type is now inferred), with a `@Deprecated` replacement hint. Rationale in [ADR 66](doc/architecture/decisions/0066-kotlin-first-dcb-criteria-construction.md).
